api/requirements.txt
fastapi
uvicorn[standard]
python-multipart
aiofiles

numpy
scipy
librosa
soundfile
matplotlib
sqlalchemy

python-dotenv
pydantic
fastapi-users[sqlalchemy]
aiosqlite
python-multipart
psycopg2
typing
pyjwt
python-jose[cryptography]
pydantic-settings
passlib>=1.7.4
cryptography>=42.0.0
